Hien: fix errors and redesign the code so that we do not have to copy back and forth between PlaygameScence.cs and ShipWreckScene.cs

+++ Poseidon/Poseidon/Enemy.cs
@@ -19,7 +19,7 @@ namespace Poseidon {
 
         //stunned and cannot move
         public bool stunned;
-        public float stunnedStartTime;
+        public double stunnedStartTime;
 
         // Time stampt since the robot starts chasing
         private TimeSpan startChasingTime;
